Add a distributed tree struct/interface
