+++ crates/rusty-tree/src/lib.rs
@@ -1,8 +1,8 @@
 pub mod adaptive;
 pub mod c_api;
 pub mod constants;
-pub mod distribute;
+pub mod distributed;
 pub mod helpers;
-pub mod octree;
+pub mod single_node;
 pub mod types;
 pub mod regular;
